ep06-simple-app/bower_components/web-animations-next/test/blink/out-of-order-keyframes.html

<!--
 Copyright 2017 Google Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
      http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
-->

<!doctype html>
<title>Web Animations API: Out of order Keyframe tests</title>
<script src="testharness/testharness.js"></script>
<script src="testharness/testharnessreport.js"></script>
<div id='log'></div>
<div id='div'></div>
<script>

test(function() {
  assert_throws({name: 'InvalidModificationError'}, function() {
    div.animate([{height: '100px', offset: 0.5}, {height: '200px', offset: 0}, {height: '300px', offset: 1}], 1);
  }, 'Out-of-order properties should not work');

  assert_throws({name: 'InvalidModificationError'}, function() {
    div.animate([{height: '100px', offset: 0.5}, {height: '150px'},
                 {height: '200px', offset: 0}, {height: '300px', offset: 1}], 1);
  }, 'Auto offsets combined with out-of-order properties should not work');
},
'Out-of-order Keyframe tests',
{
  help:   'http://dev.w3.org/fxtf/web-animations/#normalizing-a-sequence-of-keyframes',
  assert: 'Keyframes with out-of-order offsets should work, unless there are also auto offsets',
  author: 'Shane Stephens'
});

</script>
